+++ app/containers/App/actions.js
@@ -2,7 +2,7 @@
  * App Actions
  *
  * Actions change things in your application
- * Since this boilerplate uses a uni-directional data flow, specifically redux,
+ * Since this boilerplate uses a uni-directional addedQuotes flow, specifically redux,
  * we have these actions which are the only way your application interacts with
  * your application state. This guarantees that your state is up to date and nobody
  * messes it up weirdly somewhere.
@@ -35,7 +35,7 @@ export function loadRepos() {
 /**
  * Dispatched when the repositories are loaded by the request saga
  *
- * @param  {array} repos The repository data
+ * @param  {array} repos The repository addedQuotes
  * @param  {string} username The current username
  *
  * @return {object}      An action object with a type of LOAD_REPOS_SUCCESS passing the repos