+++ app/containers/QuotePage/addRowToSelected.js
@@ -1,12 +1,12 @@
 export default function addRowToSelected(currentID,selectedRows){
 
   if(selectedRows.length < 1){
-    return [currentID];
+    return [parseInt(currentID)];
   }else if(!selectedRows.includes(currentID)){
-    selectedRows.push(currentID);
+    selectedRows.push(parseInt(currentID));
     return selectedRows;
   }else{
-    selectedRows.splice(selectedRows.indexOf(currentID),1);
+    selectedRows.splice(selectedRows.indexOf(parseInt(currentID)),1);
     return selectedRows;
   }
 }