+++ app/containers/QuotePage/saga.js
@@ -1,6 +1,15 @@
 // import { take, call, put, select } from 'redux-saga/effects';
-import {call, put,takeEvery ,all,takeLatest} from 'redux-saga/effects';
-import {receiveQuote, addButtonEnable, addButtonDisable,deleteSelectedQuotes,deleteButtonEnable, deleteButtonDisable} from './actions';
+import {call, put,takeEvery ,all} from 'redux-saga/effects';
+import {
+  receiveQuote,
+  addButtonEnable,
+  addButtonDisable,
+  deleteButtonEnable,
+  deleteButtonDisable,
+  moveSelectedQuotesToDeleted,
+  removeSelectedQuotesFromData,
+  clearSelectedQuotes
+} from './actions';
 import {DELETE_BUTTON_CLICKED,ADD_BUTTON_CLICKED} from './constants';
 import axios from 'axios';
 import convertQuoteToJSON from "./convertQuoteToJSON";
@@ -10,10 +19,12 @@ function* fetchQuote(){
   yield put(addButtonDisable());
   try {
     const receivedData = yield call([axios, axios.get], 'http://ne-dev-pegasus-quotes.azurewebsites.net/api/values')
-    yield put(receiveQuote(convertQuoteToJSON(receivedData.data)));
+    const dataToJSON = convertQuoteToJSON(receivedData.data);
+    yield put(receiveQuote(dataToJSON));
     yield put(addButtonEnable());
   }catch(e){
     alert('Add quote failed. try again later');
+    console.log(e);
     yield put(addButtonEnable());
   }
 }
@@ -25,7 +36,9 @@ export function* fetchQuoteWatcher(){
 
 function* deleteButtonClicked(){
   yield put(deleteButtonDisable());
-  yield put(deleteSelectedQuotes());
+  yield put(moveSelectedQuotesToDeleted());
+  yield put(removeSelectedQuotesFromData());
+  yield put(clearSelectedQuotes());
   yield put(deleteButtonEnable());
 }
 function* deleteButtonClickedWatcher(){