+++ app/utils/request.js
@@ -37,7 +37,7 @@ function checkStatus(response) {
  * @param  {string} url       The URL we want to request
  * @param  {object} [options] The options we want to pass to "fetch"
  *
- * @return {object}           The response data
+ * @return {object}           The response addedQuotes
  */
 export default function request(url, options) {
   return fetch(url, options)